requirements.txt
loguru~=0.7.2
psycopg2-binary~=2.9.9
python-binance~=1.0.19
python-dotenv~=1.0.1